requirements.txt
langchain
python-dotenv
ipykernel
pypdf
arxiv
pymupdf
wikipedia
langchain-text-splitters
chromadb
sentence_transformers
langchain-openai
langchain_community
beautifulsoup4
faiss-cpu
langchain_groq
langchain_core
duckdb
pandas
duckduckgo-search
fastapi
uvicorn
langserve
langchain_huggingface
langchain_chroma
ipywidgets
openai